+++ rebound/tests/test_simulation.py
@@ -221,13 +221,13 @@ class TestSimulation(unittest.TestCase):
 
 
     def test_nofile(self):
-        with self.assertRaises(ValueError):
-            sim2 = rebound.Simulation.from_file("doesnotexist.bin")
+        with self.assertRaises(RuntimeError):
+            sim2 = rebound.Simulation("doesnotexist.bin")
 
 
     def test_checkpoint(self):
         self.sim.save("bintest.bin")
-        sim2 = rebound.Simulation.from_file("bintest.bin")
+        sim2 = rebound.Simulation("bintest.bin")
         self.assertEqual(self.sim.particles[1].x, sim2.particles[1].x)
         self.assertEqual(self.sim.particles[1].vx, sim2.particles[1].vx)
         self.assertEqual(self.sim.t, sim2.t)
@@ -239,7 +239,7 @@ class TestSimulation(unittest.TestCase):
         self.sim.integrate(1.)
         self.sim.save("bintest.bin")
         self.sim.integrate(5.)
-        sim2 = rebound.Simulation.from_file("bintest.bin")
+        sim2 = rebound.Simulation("bintest.bin")
         sim2.integrate(5.)
         self.assertEqual(self.sim.particles[1].x, sim2.particles[1].x)
         self.assertEqual(self.sim.particles[1].vx, sim2.particles[1].vx)