+++ src/simulationarchive.h
@@ -30,10 +30,8 @@
 struct reb_simulation;
 struct reb_particles;
 
-/**
- * @brief Internal function to handle outputs for the Simulation Archive.
- */
-void reb_simulationarchive_heartbeat(struct reb_simulation* const r);
+void reb_simulationarchive_heartbeat(struct reb_simulation* const r);  ///< Internal function to handle outputs for the Simulation Archive.
+void reb_read_simulationarchive_with_messages(struct reb_simulationarchive* sa, const char* filename, enum reb_input_binary_messages* warnings); ///< Internal function to read one snapshot from a simulation archive.
 
 
 #endif 	// SIMULATIONARCHIVE_H